requirements.txt
matplotlib==3.5.2
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.1.1
pywebio==1.6.1
torch==1.11.0
torchvision==0.12.0
